+++ packages/snap-store-mobx/src/Finder/FinderStore.ts
@@ -6,17 +6,18 @@ import { PaginationStore } from '../Search/Stores';
 import { StorageStore, StorageType } from '../Storage/StorageStore';
 import { SelectionStore } from './Stores';
 import type { FinderStoreConfig, StoreServices, SelectedSelection, FinderStoreState } from '../types';
+import { UrlManager } from '@searchspring/snap-url-manager';
 
 export class FinderStore extends AbstractStore {
-	services: StoreServices;
-	config: FinderStoreConfig;
-	data: SearchResponseModel & { meta: MetaResponseModel };
-	meta: MetaResponseModel = {};
-	storage: StorageStore;
-	persistedStorage: StorageStore;
-	pagination: PaginationStore;
-	selections: SelectionStore;
-	state: FinderStoreState = {
+	public services: StoreServices;
+	public config!: FinderStoreConfig;
+	public data!: SearchResponseModel & { meta: MetaResponseModel };
+	public meta: MetaResponseModel = {};
+	public storage: StorageStore;
+	public persistedStorage!: StorageStore;
+	public pagination!: PaginationStore;
+	public selections!: SelectionStore;
+	public state: FinderStoreState = {
 		persisted: false,
 	};
 
@@ -46,14 +47,14 @@ export class FinderStore extends AbstractStore {
 		});
 	}
 
-	setService(name, service): void {
+	public setService(name: keyof StoreServices, service: UrlManager): void {
 		if (this.services[name] && service) {
 			this.services[name] = service;
 		}
 	}
 
-	save(): void {
-		if (this.config.persist?.enabled && this.selections.filter((selection) => selection.selected).length) {
+	public save(): void {
+		if (this.config.persist?.enabled && this.persistedStorage && this?.selections?.filter((selection) => selection.selected).length) {
 			this.persistedStorage.set('config', this.config);
 			this.persistedStorage.set('data', this.data);
 			this.persistedStorage.set('date', Date.now());
@@ -70,9 +71,9 @@ export class FinderStore extends AbstractStore {
 		}
 	}
 
-	reset = (): void => {
+	public reset = (): void => {
 		if (this.config.persist?.enabled) {
-			this.persistedStorage.clear();
+			this.persistedStorage?.clear();
 			this.state.persisted = false;
 		}
 
@@ -83,15 +84,15 @@ export class FinderStore extends AbstractStore {
 		}
 	};
 
-	loadPersisted(): void {
-		if (this.config.persist?.enabled && !this.loaded) {
+	public loadPersisted(): void {
+		if (this.config.persist?.enabled && this.persistedStorage && !this.loaded) {
 			const date = this.persistedStorage.get('date');
 			const data = this.persistedStorage.get('data');
 			const config = this.persistedStorage.get('config');
 			const selections = this.persistedStorage.get('selections');
 			const isExpired = this.config.persist.expiration && Date.now() - date > this.config.persist.expiration;
 
-			if (data && selections.filter((selection) => selection.selected).length) {
+			if (data && selections.filter((selection: SelectedSelection) => selection.selected).length) {
 				if (JSON.stringify(config) === JSON.stringify(this.config) && !isExpired) {
 					this.update(data, selections);
 					this.state.persisted = true;
@@ -103,7 +104,7 @@ export class FinderStore extends AbstractStore {
 		}
 	}
 
-	update(data: SearchResponseModel & { meta: MetaResponseModel }, selectedSelections?: SelectedSelection[]): void {
+	public update(data: SearchResponseModel & { meta: MetaResponseModel }, selectedSelections?: SelectedSelection[]): void {
 		this.error = undefined;
 		this.data = JSON.parse(JSON.stringify(data));
 		this.loaded = !!data.pagination;
@@ -111,11 +112,11 @@ export class FinderStore extends AbstractStore {
 		this.pagination = new PaginationStore(this.config, this.services, data.pagination);
 		this.selections = new SelectionStore(this.config, this.services, {
 			state: this.state,
-			facets: data.facets,
+			facets: data.facets || [],
 			meta: this.meta,
 			loading: this.loading,
 			storage: this.storage,
-			selections: selectedSelections,
+			selections: selectedSelections || [],
 		});
 	}
 }