+++ packages/snap-store-mobx/src/Recommendation/Stores/ProfileStore.ts
@@ -1,19 +1,20 @@
 import { observable, makeObservable } from 'mobx';
 import type { StoreServices } from '../../types';
+import type { RecommendCombinedResponseModel } from '@searchspring/snap-client';
 
 export class ProfileStore {
-	tag: string;
-	placement: string;
-	display: Record<string, any> = {};
+	public tag!: string;
+	public placement!: string;
+	public display: Record<string, any> = {};
 
-	constructor(services: StoreServices, profile) {
-		if (!profile?.tag) {
+	constructor(services: StoreServices, data?: RecommendCombinedResponseModel) {
+		if (!data?.profile?.tag) {
 			return;
 		}
 
-		this.tag = profile.tag;
-		this.placement = profile.placement;
-		this.display = profile.display;
+		this.tag = data.profile.tag;
+		this.placement = data.profile.placement;
+		this.display = data.profile.display;
 
 		makeObservable(this, {
 			tag: observable,