+++ packages/snap-store-mobx/src/Search/SearchStore.test.ts
@@ -2,6 +2,7 @@ import { UrlManager, UrlTranslator } from '@searchspring/snap-url-manager';
 import { MockData } from '@searchspring/snap-shared';
 
 import { SearchStore } from './SearchStore';
+import { SearchResponseModel, MetaResponseModel } from '@searchspring/snapi-types';
 
 const services = {
 	urlManager: new UrlManager(new UrlTranslator()),
@@ -14,7 +15,7 @@ const searchConfig = {
 const mockData = new MockData();
 
 describe('Search Store', () => {
-	let searchData;
+	let searchData: SearchResponseModel & { meta: MetaResponseModel };
 
 	beforeEach(() => {
 		searchData = mockData.searchMeta();
@@ -44,10 +45,10 @@ describe('Search Store', () => {
 		expect(searchStore.results).toHaveLength(0);
 
 		expect(searchStore.pagination).toBeDefined();
-		expect(searchStore.pagination.totalResults).toBeUndefined();
+		expect(searchStore.pagination?.totalResults).toBeUndefined();
 
 		expect(searchStore.sorting).toBeDefined();
-		expect(searchStore.sorting.options).toHaveLength(0);
+		expect(searchStore.sorting?.options).toHaveLength(0);
 	});
 
 	it('update function updates all of the stores', () => {
@@ -58,19 +59,19 @@ describe('Search Store', () => {
 		expect(searchStore.meta).toStrictEqual(searchData.meta);
 
 		expect(searchStore.search).toBeDefined();
-		expect(searchStore.search.query).toStrictEqual(searchData.search.query);
+		expect(searchStore.search?.query).toStrictEqual(searchData.search?.query);
 
 		expect(searchStore.merchandising).toBeDefined();
 		expect(searchStore.merchandising).toEqual(searchData.merchandising);
 
-		expect(searchStore.facets).toHaveLength(searchData.facets.length);
+		expect(searchStore.facets).toHaveLength(searchData.facets?.length!);
 
 		expect(searchStore.filters).toHaveLength(0);
 
-		expect(searchStore.results).toHaveLength(searchData.results.length);
+		expect(searchStore.results).toHaveLength(searchData.results?.length!);
 
-		expect(searchStore.pagination.totalResults).toBe(searchData.pagination.totalResults);
+		expect(searchStore.pagination?.totalResults).toBe(searchData.pagination?.totalResults);
 
-		expect(searchStore.sorting.options).toHaveLength(searchData.meta.sortOptions.filter((option) => option.type == 'field').length);
+		expect(searchStore.sorting?.options).toHaveLength(searchData.meta.sortOptions?.filter((option) => option.type == 'field').length!);
 	});
 });