+++ packages/snap-store-mobx/src/types.ts
@@ -25,7 +25,7 @@ export type SearchStoreConfig = StoreConfig & {
 	};
 };
 
-type FacetStoreConfig = {
+export type FacetStoreConfig = {
 	trim?: boolean;
 	pinFiltered?: boolean;
 	storeRange?: boolean;
@@ -76,12 +76,13 @@ export type RecommendationStoreConfig = StoreConfig & {
 	branch?: string;
 	realtime?: boolean;
 	batched?: boolean;
+	order?: number;
 };
 
 export type StoreConfigs = SearchStoreConfig | AutocompleteStoreConfig | FinderStoreConfig | RecommendationStoreConfig;
 
 export type StoreServices = {
-	urlManager?: UrlManager;
+	urlManager: UrlManager;
 };
 
 export enum ErrorType {